+++ test/vajra_test.dart
@@ -5,7 +5,7 @@ import 'package:vajra/vajra.dart';
 void main() {
 
   test("geting set-cookie header format", () async {
-    Vajra client = Vajra("<directory.path>",basePath: "http://127.0.0.1:8000");
+    Vajra client = Vajra("test", "<directory.path>",basePath: "http://127.0.0.1:8000");
     // await client.initialize();
     final VajraResponse response = await client.get("/setcookie");
     expect(response.statusCode, 200);