+++ CHANGELOG.md
@@ -9,3 +9,10 @@
 ## 0.0.3
 
 - Resolve bug while importing security schemes.
+
+## 0.0.4
+
+- resolved bug while attaching multiple queries.
+- started supporting multiple instances of vajra client using name.
+- added timeout to every request.
+- added a method for setting default timeout.